requirements.txt
streamlit
langchain
python-dotenv
openai
pandas
tabulate